ENH: Refactor fieldmap-unwarping flows, more homogeneous interface
This PR closes #19, reorganizing the unwarping tools for fieldmap-based
solutions:

  - [x] Separate the displacements field generation from actual
        unwarping (i.e., applying the nonlinear transform).
        By doing this, we are not only addressing #19 directly, we are
        also readying the ground for tackling #21 since now all the
        unwarping paths (pepolar, fieldmaps/phases/phasediff, or syn)
        have more consistent interfaces.
  - [x] Update the root workflow to reflect these changes.
  - [x] Unloaded ``unwarp.py`` of the workflow to write reports, which
        has been moved to a new module called ``outputs.py`` following
        some of the latest *fMRIPrep* non-written standards.
  - [x] Polished some minor documentation and stylistic issues within
        the scope of the proposed changes.
